Refactor/ NavBar 분리
